chore(Refactor API route functions): refactored API routing function into controller class
