+++ src/main/app.js
@@ -27,6 +27,7 @@ app.use(`${apiVersion}/entries`, entriesRouter);
 app.use((req, res, next) => {
   next(createError(404));
 });
+
 // error handler
 app.use((err, req, res) => {
   // set locals, only providing error in development
